Deduplicate messagesMap
JIRA: RAIL-4273
